+++ test/n_security_token_registry.js
@@ -1119,6 +1119,9 @@ contract("SecurityTokenRegistry", async (accounts) => {
             let data = await I_Getter.getSecurityTokenData.call(I_SecurityToken.address);
             assert.equal(data[0], symbol);
             assert.equal(data[1], token_owner);
+            assert.equal(data[4][0], 2);
+            assert.equal(data[4][1], 0);
+            assert.equal(data[4][2], 0);
         });
 
         it("Should get the tickers by owner", async () => {